Move animation logic to custom hook
