+++ packages/harpguru-core/src/components/option-value/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useScaleOnUpdateAnimation } from './use-scale-on-update-animation'
